Move implementation of FixedSizeButton
